first changes to the branch

+++ cytograph/pipeline/commands.py
@@ -324,7 +324,8 @@ def qc(sampleids: List[str], rerun: bool = False, file: str = None, fixed_thresh
 @cli.command()
 @click.option('--subset', default=None)
 @click.option('--method', default='coverage', type=click.Choice(['coverage', 'dendrogram', 'cluster']))
-def split(subset: str = None, method: str = 'coverage') -> None:
+@click.option('--hierarchy', default='agg', type=click.Choice(['agg', 'paris']))
+def split(subset: str = None, method: str = 'coverage', hierarchy: str = 'agg') -> None:
 
     config = load_config()
     exportdir = os.path.abspath(os.path.join(config.paths.build, "exported"))
@@ -332,7 +333,7 @@ def split(subset: str = None, method: str = 'coverage') -> None:
     if subset:
 
         logging.info(f"Splitting {subset}...")
-        if split_subset(config, subset, method):
+        if split_subset(config, subset, method, hierarchy):
             deck = PunchcardDeck(config.paths.build)
             card = deck.get_card(subset)
             Workflow(deck, "").compute_subsets(card)
@@ -378,7 +379,7 @@ def split(subset: str = None, method: str = 'coverage') -> None:
 
                     # get command for task
                     task = subset.longname()
-                    cmd = f"split --subset {task} --method {method}"
+                    cmd = f"split --subset {task} --method {method} --hierarchy {hierarchy}"
 
                     # create submit file for split
                     with open(os.path.join(exdir, task + ".condor"), "w") as f:
@@ -465,20 +466,23 @@ def merge(subset: str = None, overwrite: bool = False) -> None:
         datadir = os.path.join(config.paths.build, "data")
         exportdir = os.path.join(config.paths.build, "exported")
         if not overwrite:
-            logging.info("Rearranging directories...")
+            logging.info("Backing up data...")
             shutil.copytree(datadir, os.path.join(config.paths.build, "data_premerge"))
-            shutil.copytree(exportdir, os.path.join(config.paths.build, "exported_premerge"))
+        logging.info("Backing up export...")
+        shutil.copytree(exportdir, os.path.join(config.paths.build, "exported_premerge"))
 
         logging.info("Submitting jobs")
         for subset in deck.get_leaves():
             # Use CPUs and memory from subset config
             config = load_config(subset)
-            n_cpus = config.execution.n_cpus
+            n_cpus = min(config.execution.n_cpus, 14)
             memory = config.execution.memory
             # Remove agg file and export directory
             task = subset.longname()
-            os.remove(os.path.join(datadir, task + ".agg.loom"))
-            shutil.rmtree(os.path.join(exportdir, task))
+            if os.path.exists(os.path.join(datadir, task + ".agg.loom")):
+                os.remove(os.path.join(datadir, task + ".agg.loom"))
+            if os.path.exists(os.path.join(exportdir, task)):
+                shutil.rmtree(os.path.join(exportdir, task))
             # Make submit file
             cmd = f"merge --subset {task}"
             with open(os.path.join(exdir, task + ".condor"), "w") as f:
